// frontend/src/types.ts
export interface CustomInvoiceItem {
  name: string;
  amount: number;
}

export interface CustomInvoice {
  id?: string;
  company_id: string;
  company_name?: string;
  client_name?: string;
  description: string;
  amount: number;
  due_date: string;
  invoice_month?: string;
}

export interface InvoiceQuantity {
  quantity: number;
  price: number;
}

export interface Invoice {
  id: string;
  company_id: string;
  company_name?: string;
  email?: string;
  invoice_month?: string;
  owner_name?: string;
  amount: number;
  invoice_value?: number;
  status: 'Paid' | 'Pending' | 'Failed' | 'Upcoming';
  invoice_submitted?: boolean;
  invoice_submitted_admin?: boolean;
  due_date: string;
  paid_date?: string;
  documents_accessed: number;
  storage_assigned: number;
  created_at?: string;
  document_uploaded?: number;
  document_shared?: number;
  document_downloaded?: number;
  upload_amount?: number;
  shared_amount?: number;
  download_amount?: number;
  monthly?: number;
  total?: number;
  other_invoices?: CustomInvoiceItem[];
}